+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/Serialization/Formatters/Enum/OnCompletionFormatter.cs
@@ -1,5 +1,4 @@
 using AlgoSdk.MessagePack;
-using Unity.Collections;
 
 namespace AlgoSdk.Formatters
 {