+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/WalletConnect/WebSocket/WebSocketEvent.cs
@@ -1,0 +1,22 @@
+using System;
+
+namespace Netcode.Transports.WebSocket
+{
+    public class WebSocketEvent
+    {
+        public enum WebSocketEventType
+        {
+            Nothing,
+            Open,
+            Close,
+            Payload,
+            Error,
+        }
+
+        public WebSocketEventType Type;
+        public ulong ClientId;
+        public byte[] Payload;
+        public string Error;
+        public string Reason;
+    }
+}